fix (dev-server) move helpers back to dev-server
