+++ pkg/platform/dev-server.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"os"
 	"os/signal"
+	"path/filepath"
 	"syscall"
 	"time"
 
@@ -19,26 +20,35 @@ type Server struct {
 	reload (chan bool)
 }
 
-// Start runs the dev server with the given filesystem in localhost:3049
-func (s *Server) Start(files filesystem.FileSystem) {
-	// HotReload
-	go http.HandleFunc("/hmr", s.SendBundleUpdates)
+// ContentTypes defines the supported file content types
+var ContentTypes = map[string]string{
+	".css": "text/css; charset=UTF-8",
 
-	// FileServer
-	fileServer := FileServer(http.Dir("__build__"), files)
-	go http.Handle("/", http.StripPrefix("/", fileServer))
+	".js":  "application/javascript; charset=UTF-8",
+	".mjs": "application/javascript; charset=UTF-8",
 
-	fmt.Printf("Available on http://127.0.0.1:%d\n", 3049)
-	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%d", 3049), nil))
+	".json":   "application/json; charset=UTF-8",
+	".jsonld": "application/ld+json; charset=UTF-8",
+
+	".png":  "image/png",
+	".webp": "image/webp",
+	".jpg":  "image/jpeg",
+	".jpeg": "image/jpeg",
+	".svg":  "image/svg+xml; charset=utf-8",
+
+	".woff":  "font/woff",
+	".woff2": "font/woff2",
 }
 
-// GracefullShutdown gracefully shuts down the platform bein executed
-func GracefullShutdown() {
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
+// SetContentType adds the given file's content type to the header
+func SetContentType(w http.ResponseWriter, file string) {
+	ext := filepath.Ext(file)
+	w.Header().Add("Content-Type", ContentTypes[ext])
+}
 
-	<-c
-	os.Exit(1)
+// SetCacheDuration adds Cache-Control header with the given amount of seconds
+func SetCacheDuration(w http.ResponseWriter, seconds int64) {
+	w.Header().Add("Cache-Control", fmt.Sprintf("max-age=%d", seconds))
 }
 
 type UpdateMessage struct {
@@ -77,4 +87,26 @@ func (s *Server) SendBundleUpdates(w http.ResponseWriter, r *http.Request) {
 			}
 		}
 	}
+}
+
+// Start runs the dev server with the given filesystem in localhost:3049
+func (s *Server) Start(files filesystem.FileSystem) {
+	// HotReload
+	go http.HandleFunc("/hmr", s.SendBundleUpdates)
+
+	// FileServer
+	fileServer := FileServer(http.Dir("__build__"), files)
+	go http.Handle("/", http.StripPrefix("/", fileServer))
+
+	fmt.Printf("Available on http://127.0.0.1:%d\n", 3049)
+	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%d", 3049), nil))
+}
+
+// GracefullShutdown gracefully shuts down the platform bein executed
+func GracefullShutdown() {
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
+
+	<-c
+	os.Exit(1)
 }
